requirements.txt
dnspython>=2.4.0
pyyaml>=6.0
tabulate>=0.9.0
